requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2020.6.20
dj-database-url==0.5.0
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
numpy==1.22.3
pandas==1.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.0.0
